requirements.txt
asgiref==3.5.2
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
Pillow==9.2.0
pytz==2022.2.1
sqlparse==0.4.2
